requirements.txt
requests==2.31.0
pydantic==2.3.0
fastapi==0.103.0
uvicorn==0.23.2
psutil>=5.8.0
puts==0.0.11
supabase
rich
python-dotenv